// packages/api/src/api/types/generated.ts
export type ConfigMap = Record<string, string>;

/**
 * DevicesMap type is used to hold incus devices configurations. In contrast to
 * plain map[string]map[string]string it provides unmarshal methods for JSON and
 * YAML, which gracefully handle numbers and bools.
 *
 *
 * @example {"eth0": {"network": "incusbr0", "type": "nic"}}
 */
export type DevicesMap = Record<string, Record<string, string>>;
// ...
export interface Network extends NetworkPut {
  /**
   * The network name
   * Read only: true
   * @example mybr0
   */
  name: string;
  /**
   * The network type
   * Read only: true
   * @example bridge
   */
  type: string;
  /**
   * List of URLs of objects using this profile
   * Read only: true
   * @example ["/1.0/profiles/default", "/1.0/instances/c1"]
   */
  used_by: string[];
  /**
   * Whether this is a managed network
   * Read only: true
   * @example true
   *
   * API extension: network
   */
  managed: boolean;
  /**
   * The state of the network (for managed network in clusters)
   * Read only: true
   * @example Created
   *
   * API extension: clustering
   */
  status: string;
  /**
   * Cluster members on which the network has been defined
   * Read only: true
   * @example ["server01", "server02", "server03"]
   *
   * API extension: clustering
   */
  locations: string[];
  /**
   * Project name
   * @example project1
   *
   * API extension: networks_all_projects
   */
  project: string;
}
// ...
export interface NetworkPut {
  /**
   * Network configuration map (refer to doc/networks.md)
   * @example {"ipv4.address": "10.0.0.1/24", "ipv4.nat": "true", "ipv6.address": "none"}
   */
  config: ConfigMap;
  /**
   * Description of the profile
   * @example My new bridge
   *
   * API extension: entity_description
   */
  description: string;
}
// ...
export interface Profile extends ProfilePut {
  /**
   * The profile name
   * Read only: true
   * @example foo
   */
  name: string;
  /**
   * List of URLs of objects using this profile
   * Read only: true
   * @example ["/1.0/instances/c1", "/1.0/instances/v1"]
   *
   * API extension: profile_usedby
   */
  used_by: string[];
  /**
   * Project name
   * @example project1
   *
   * API extension: profiles_all_projects
   */
  project: string;
}

/**
 * ProfilePost represents the fields required to rename a profile
 */
export interface ProfilePost {
  /**
   * The new name for the profile
   * @example bar
   */
  name: string;
}

/**
 * ProfilePut represents the modifiable fields of a profile
 */
export interface ProfilePut {
  /**
   * Instance configuration map (refer to doc/instances.md)
   * @example {"limits.cpu": "4", "limits.memory": "4GiB"}
   */
  config: ConfigMap;
  /**
   * Description of the profile
   * @example Medium size instances
   */
  description: string;
  /**
   * List of devices
   * @example {"root": {"type": "disk", "pool": "default", "path": "/"}, "eth0": {"type": "nic", "network": "mybr0", "name": "eth0"}}
   */
  devices: DevicesMap;
}
// ...
export interface StoragePool extends StoragePoolPut {
  /**
   * Storage pool name
   * @example local
   */
  name: string;
  /**
   * Storage pool driver (btrfs, ceph, cephfs, cephobject, dir, lvm, lvmcluster or zfs)
   * @example zfs
   */
  driver: string;
  /**
   * List of URLs of objects using this storage pool
   * @example ["/1.0/profiles/default", "/1.0/instances/c1"]
   */
  used_by: string[];
  /**
   * Pool status (Pending, Created, Errored or Unknown)
   * Read only: true
   * @example Created
   *
   * API extension: clustering
   */
  status: string;
  /**
   * Cluster members on which the storage pool has been defined
   * Read only: true
   * @example ["server01", "server02", "server03"]
   *
   * API extension: clustering
   */
  locations: string[];
}

/**
 * StoragePoolPut represents the modifiable fields of a storage pool.
 *
 *
 * API extension: storage.
 */
export interface StoragePoolPut {
  /**
   * Storage pool configuration map (refer to doc/storage.md)
   * @example {"volume.block.filesystem": "ext4", "volume.size": "50GiB"}
   */
  config: ConfigMap;
  /**
   * Description of the storage pool
   * @example Local SSD pool
   *
   * API extension: entity_description
   */
  description: string;
}